requirements.txt
fastapi
uvicorn
pydantic
python-multipart
requests
beautifulsoup4
python-dotenv
openai
newspaper4k
lxml_html_clean